Style Exploration
First round style sketches… these aren’t pixel perfect. Just trying to
work out the build/style. Would be perfected later. That security
system is a tricky one. I have a few different ideas for what it could
be, within each style. Also did a quick take on the humidity and
motion. None of them are likely a done deal, but it’s a start.